feat: 자동차 이름 원시값 포장  - 자동차 이름의 적합성 판별 로직도 자동차 클래스에서 자동차 이름 클래스로 이동
